more tests, refactors, and networking setup
