+++ app/src/components/RobotSettings/SelectNetwork/constants.js
@@ -1,5 +1,11 @@
 // @flow
 
+export const ACTIVE = 'ACTIVE'
+
+export const CONNECT = 'connect'
+export const DISCONNECT = 'disconnect'
+export const JOIN = 'join'
+
 export const DISCONNECT_WIFI_VALUE = '__disconnect-from-wifi__'
 export const DISCONNECT_WIFI_LABEL = 'Disconnect from Wifi'
 export const JOIN_OTHER_VALUE = '__join-other-network__'
